feature(mailbox-creator): rename to creator
