+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/parsing/mailbox/creator/MailboxCreator.java
@@ -1,8 +1,8 @@
-package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator;
+package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.creator;
 
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
 
-public interface MailboxGenerator {
+public interface MailboxCreator {
 
   String generateMailboxContent(Mailbox mailbox);
 }
