+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/external/infrastructure/email/JSONMailboxCreator.java
@@ -3,9 +3,9 @@ package de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email;
 import static de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.GSONConfiguration.getConfiguredGson;
 
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.MailboxGenerator;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.creator.MailboxCreator;
 
-public class JSONMailboxGenerator implements MailboxGenerator {
+public class JSONMailboxCreator implements MailboxCreator {
 
   @Override
   public String generateMailboxContent(Mailbox mailbox) {
