+++ src/test/java/de/dhbw/karlsruhe/students/mailflow/unit/JSONParserTest.java
@@ -37,7 +37,7 @@ class JSONParserTest {
 
     Mailbox expectedMailbox = Mailbox.create(owner, List.of(), MailboxType.READ);
     Email expectedEmail = Email.create(emailContent, emailMetadata);
-    // This string was automatically generated by the MboxGenerator class.
+    // This string was automatically generated by the MboxCreator class.
     // Currently, it is not in this branch
     String jsonString =
         """
