+++ src/test/java/de/dhbw/karlsruhe/students/mailflow/unit/LocalMailboxRepositoryTest.java
@@ -4,8 +4,8 @@ import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.MailboxRe
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.enums.MailboxType;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.Address;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.FileCreationException;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.FileWritingException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.creator.FileCreationException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.creator.FileWritingException;
 import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.LocalMailboxRepository;
 import java.io.File;
 import java.util.List;
